+++ src/components/Icon/__generated__/bootstrap/Anthropic.icon.tsx
@@ -5,13 +5,13 @@ import { CommonIconProps, commonIconDefaulProps } from "../common.types";
 import { cssLengthUsage, cssVarUsage, mergeDefaultProps } from "../../../../utils";
 
 /**
- * React component generated from bootstrap icon: `unlock`.
+ * React component generated from bootstrap icon: `anthropic`.
  * Auto-generated before NPM release
  *
  * @params {CommonIconProps} [props = commonIconDefaulProps] - the icon's size and color
  * @returns {JSX.Element} - the icon
  */
-const Unlock = (props: CommonIconProps): JSX.Element => {
+const Anthropic = (props: CommonIconProps): JSX.Element => {
   const { color, size } = useMemo(() => {
     const mergedProps = mergeDefaultProps(commonIconDefaulProps, props);
     return {
@@ -31,10 +31,10 @@ const Unlock = (props: CommonIconProps): JSX.Element => {
     >
       <path
         fillRule="evenodd"
-        d="M12 0a4 4 0 0 1 4 4v2.5h-1V4a3 3 0 1 0-6 0v2h.5A2.5 2.5 0 0 1 12 8.5v5A2.5 2.5 0 0 1 9.5 16h-7A2.5 2.5 0 0 1 0 13.5v-5A2.5 2.5 0 0 1 2.5 6H8V4a4 4 0 0 1 4-4M2.5 7A1.5 1.5 0 0 0 1 8.5v5A1.5 1.5 0 0 0 2.5 15h7a1.5 1.5 0 0 0 1.5-1.5v-5A1.5 1.5 0 0 0 9.5 7z"
+        d="M9.218 2h2.402L16 12.987h-2.402zM4.379 2h2.512l4.38 10.987H8.82l-.895-2.308h-4.58l-.896 2.307H0L4.38 2.001zm2.755 6.64L5.635 4.777 4.137 8.64z"
       />
     </svg>
   );
 };
 
-export default Unlock;
+export default Anthropic;